requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pyyaml==6.0.1
playwright>=1.49.0
pytz==2024.1
schedule==1.2.1
fake-useragent==1.5.1
python-dotenv==1.0.1
